\name{svd_positive}
\alias{svd_positive}
\title{Singular Value Decomposition (SVD) retaining positive values}
\usage{
  svd_positive(m, tol)
}
\arguments{
  \item{m}{a numeric matrix}

  \item{tol}{tolerance threshold for positiveness of
  singular values}
}
\description{
  Calculates SVD that retains only positive singular values
}
\keyword{internal}

