\name{svd_tall}
\alias{svd_tall}
\title{Singular Value Decomposition (SVD) for a tall matrix}
\usage{
  svd_tall(m, tol)
}
\arguments{
  \item{m}{a numeric matrix}

  \item{tol}{tolerance threshold for singular values}
}
\description{
  SVD of tall matrix (number of rows >> number of columns)
}
\keyword{internal}

