\name{svd_wide}
\alias{svd_wide}
\title{Singular Value Decomposition (SVD) for a wide matrix}
\usage{
  svd_wide(m, tol)
}
\arguments{
  \item{m}{a numeric matrix}

  \item{tol}{tolerance threshold for singular values}
}
\description{
  SVD of wide matrix (number of rows << number of columns)
}
\keyword{internal}

