\name{weighted_average}
\alias{weighted_average}
\title{Weighted Average}
\usage{
  weighted_average(v, weights = NULL)
}
\arguments{
  \item{v}{numeric vector}

  \item{weights}{optional numeric vector}
}
\description{
  Computes a weighted average of the elements in a vector
}
\keyword{internal}

